Move page container and sidebar navigation to separate layout folder
