+++ features/ui/index.ts
@@ -1,4 +1,2 @@
 export * from "./badge";
 export * from "./button";
-export * from "./page-container";
-export * from "./sidebar-navigation";
