+++ pages/_app.tsx
@@ -5,7 +5,7 @@ import type { AppProps } from "next/app";
 import { ThemeProvider } from "styled-components";
 import { QueryClientProvider } from "@tanstack/react-query";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
-import { NavigationProvider } from "@features/ui";
+import { NavigationProvider } from "@features/layout";
 import { GlobalStyle } from "@styles/global-style";
 import { theme } from "@styles/theme";
 import { queryClient } from "@api/query-client";
