+++ pages/dashboard/issues.tsx
@@ -1,4 +1,4 @@
-import { PageContainer } from "@features/ui";
+import { PageContainer } from "@features/layout";
 import { IssueList } from "@features/issues";
 import type { NextPage } from "next";
